Renaming packages from lab02a_jsteck to assign01
