+++ war/WEB-INF/web.xml
@@ -8,33 +8,33 @@
   <!-- Servlets -->
   <servlet>
     <servlet-name>IndexServlet</servlet-name>
-    <servlet-class>edu.ycp.cs320.lab02a_jsteck.servlet.IndexServlet</servlet-class>
+    <servlet-class>edu.ycp.cs320.assign01.servlet.IndexServlet</servlet-class>
   </servlet>
   <servlet>
     <servlet-name>AddNumbersServlet</servlet-name>
-    <servlet-class>edu.ycp.cs320.lab02a_jsteck.servlet.AddNumbersServlet</servlet-class>
+    <servlet-class>edu.ycp.cs320.assign01.servlet.AddNumbersServlet</servlet-class>
   </servlet>
   <servlet>
     <servlet-name>MultiplyNumbersServlet</servlet-name>
-    <servlet-class>edu.ycp.cs320.lab02a_jsteck.servlet.MultiplyNumbersServlet</servlet-class>
+    <servlet-class>edu.ycp.cs320.assign01.servlet.MultiplyNumbersServlet</servlet-class>
   </servlet>
   <servlet>
     <servlet-name>GuessingGameServlet</servlet-name>
-    <servlet-class>edu.ycp.cs320.lab02a_jsteck.servlet.GuessingGameServlet</servlet-class>
+    <servlet-class>edu.ycp.cs320.assign01.servlet.GuessingGameServlet</servlet-class>
   </servlet>
   
   <!-- AJAX servlets. -->
   <servlet>
     <servlet-name>AddNumbersAjaxServlet</servlet-name>
-    <servlet-class>edu.ycp.cs320.lab02a_jsteck.servlet.ajax.AddNumbersAjaxServlet</servlet-class>
+    <servlet-class>edu.ycp.cs320.assign01.servlet.ajax.AddNumbersAjaxServlet</servlet-class>
   </servlet>
   <servlet>
     <servlet-name>MultiplyNumbersAjaxServlet</servlet-name>
-    <servlet-class>edu.ycp.cs320.lab02a_jsteck.servlet.ajax.MultiplyNumbersAjaxServlet</servlet-class>
+    <servlet-class>edu.ycp.cs320.assign01.servlet.ajax.MultiplyNumbersAjaxServlet</servlet-class>
   </servlet>
   <servlet>
     <servlet-name>NextGuessAjaxServlet</servlet-name>
-    <servlet-class>edu.ycp.cs320.lab02a_jsteck.servlet.ajax.NextGuessAjaxServlet</servlet-class>
+    <servlet-class>edu.ycp.cs320.assign01.servlet.ajax.NextGuessAjaxServlet</servlet-class>
   </servlet>
 
   <!-- Servlet mappings -->
